convert to bigint

+++ priv/repo/migrations/20241121180120_create_accounting.exs
@@ -3,7 +3,7 @@ defmodule Airlink.Repo.Migrations.CreateAccounting do
 
   def change do
     create table(:accounting) do
-      add :acct_delay_time, :integer
+      add :acct_delay_time, :bigint
       add :acct_session_id, :string
       add :acct_status_type, :string
       add :acct_unique_session_id, :string, null: false
@@ -16,13 +16,13 @@ defmodule Airlink.Repo.Migrations.CreateAccounting do
       add :nas_port_id, :string
       add :nas_port_type, :string
       add :user_name, :string
-      add :acct_input_gigawords, :integer
-      add :acct_input_octets, :integer
-      add :acct_input_packets, :integer
-      add :acct_output_gigawords, :integer
-      add :acct_output_octets, :integer
-      add :acct_output_packets, :integer
-      add :acct_session_time, :integer
+      add :acct_input_gigawords, :bigint
+      add :acct_input_octets, :bigint
+      add :acct_input_packets, :bigint
+      add :acct_output_gigawords, :bigint
+      add :acct_output_octets, :bigint
+      add :acct_output_packets, :bigint
+      add :acct_session_time, :bigint
       add :subscription_id, :uuid
       add :acct_terminate_cause, :string
       add :company_id, :uuid
